Renamed the buttons directory
