+++ src/components/bar/navigationBar/filterBar/DeleteFilterAlertDialog.jsx
@@ -11,7 +11,7 @@ import {
 } from "@chakra-ui/react";
 import { useTranslation } from "react-i18next";
 
-import ButtonComponent from "../../../basic/Buttons/ButtonComponent";
+import ButtonComponent from "../../../basic/buttons/ButtonComponent";
 import {
   greyColor,
   redColor,
