+++ src/components/bar/navigationBar/filterBar/FilterBar.jsx
@@ -14,7 +14,7 @@ import {
   whiteColor,
 } from "../../../../helper/constant";
 import FilterMenuButton from "./filtersButtons/FilterMenuButton";
-import ButtonComponent from "../../../basic/Buttons/ButtonComponent";
+import ButtonComponent from "../../../basic/buttons/ButtonComponent";
 import SaveFilterAlertDialog from "./SaveFilterAlertDialog";
 import SavedFilterButton from "./filtersButtons/SavedFilterButton";
 import DeleteFilterAlertDialog from "./DeleteFilterAlertDialog";
