+++ src/components/bar/sideBar/SideBar.jsx
@@ -7,7 +7,7 @@ import { Archive } from "@styled-icons/boxicons-regular";
 import { Accordion } from "@chakra-ui/accordion";
 import { useTranslation } from "react-i18next";
 
-import ButtonComponent from "../../basic/Buttons/ButtonComponent";
+import ButtonComponent from "../../basic/buttons/ButtonComponent";
 import AccordionItemComponent from "./AccordionItemComponent";
 import {
   blackColor,
